requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Async Support
httpx==0.25.1
aiohttp==3.9.1

# Optional local agent runtime
# Install manually if you want the free AutoGen + Ollama demo:
# pip install "pyautogen[ollama]"

# OpenTelemetry (Observability)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-httpx==0.42b0

# Configuration
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx[testing]==0.25.1

# Development & Debugging
rich==13.7.0
ipython==8.18.1
black==23.12.0
ruff==0.1.8

# Optional: Monitoring Dashboard (for development)
prometheus-client==0.19.0
